Giving birth to the best export builtin version
Adding envlst updating function
Adding extand_words

+++ src/envlst/envlst_find_var.c
@@ -6,7 +6,7 @@
 /*   By: ffeaugas <ffeaugas@student.42angouleme.fr  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 17:56:28 by ffeaugas          #+#    #+#             */
-/*   Updated: 2023/02/17 15:33:56 by tdubois          ###   ########.fr       */
+/*   Updated: 2023/02/17 16:41:21 by ffeaugas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ t_envlst	*my_envlst_find_var(char const *varspec, t_envlst *envlst)
 	}
 	while (envlst != NULL)
 	{
-		if (len == ft_strcspn(envlst->content, "=")
+		if (len == (int) ft_strcspn(envlst->content, "=")
 				&& ft_strncmp(envlst->content, varspec, len))
 			return (envlst);
 		envlst = envlst->next;
